+++ mcnn/models.py
@@ -1,6 +1,8 @@
 import torch
 import torch.nn as nn
-from network import Conv2d
+
+from mcnn.network import Conv2d
+
 
 class MCNN(nn.Module):
     '''
