+++ mcnn/network.py
