+++ mcnn/timer.py
