+++ mcnn/utils.py
