+++ test.py
@@ -1,11 +1,12 @@
+from __future__ import print_function
 import os
 import torch
 import numpy as np
 
-from src.crowd_count import CrowdCounter
-from src import network
-from src.data_loader import ImageDataLoader
-from src import utils
+from mcnn.crowd_count import CrowdCounter
+from mcnn import network
+from mcnn.data_loader import ImageDataLoader
+from mcnn import utils
 
 
 torch.backends.cudnn.enabled = True
@@ -55,7 +56,7 @@ for blob in data_loader:
         
 mae = mae/data_loader.get_num_samples()
 mse = np.sqrt(mse/data_loader.get_num_samples())
-print '\nMAE: %0.2f, MSE: %0.2f' % (mae,mse)
+print('\nMAE: %0.2f, MSE: %0.2f' % (mae,mse))
 
 f = open(file_results, 'w') 
 f.write('MAE: %0.2f, MSE: %0.2f' % (mae,mse))
